Crate rusoto_pi

AWS Performance Insights enables you to monitor and explore different dimensions of database load based on data captured from a running RDS instance. The guide provides detailed information about Performance Insights data types, parameters and errors. For more information about Performance Insights capabilities see Using Amazon RDS Performance Insights in the Amazon RDS User Guide.

The AWS Performance Insights API provides visibility into the performance of your RDS instance, when Performance Insights is enabled for supported engine types. While Amazon CloudWatch provides the authoritative source for AWS service vended monitoring metrics, AWS Performance Insights offers a domain-specific view of database load measured as Average Active Sessions and provided to API consumers as a 2-dimensional time-series dataset. The time dimension of the data provides DB load data for each time point in the queried time range, and each time point decomposes overall load in relation to the requested dimensions, such as SQL, Wait-event, User or Host, measured at that time point.

If you're using the service, you're probably looking for PerformanceInsightsClient and PerformanceInsights.

Structs

DataPoint
A timestamp, and a single numerical value, which together represent a measurement at a particular point in time.

DescribeDimensionKeysRequest

DescribeDimensionKeysResponse

DimensionGroup
A logical grouping of Performance Insights metrics for a related subject area. For example, the db.sql dimension group consists of the following dimensions: db.sql.id, db.sql.db_id, db.sql.statement, and db.sql.tokenized_id.

DimensionKeyDescription
An array of descriptions and aggregated values for each dimension within a dimension group.

GetResourceMetricsRequest

GetResourceMetricsResponse

MetricKeyDataPoints
A time-ordered series of data points, correpsonding to a dimension of a Performance Insights metric.

MetricQuery
A single query to be processed. You must provide the metric to query. If no other parameters are specified, Performance Insights returns all of the data points for that metric. You can optionally request that the data points be aggregated by dimension group ( GroupBy), and return only those data points that match your criteria (Filter).

PerformanceInsightsClient
A client for the AWS PI API.

ResponsePartitionKey
If PartitionBy was specified in a DescribeDimensionKeys request, the dimensions are returned in an array. Each element in the array specifies one dimension.

ResponseResourceMetricKey
An object describing a Performance Insights metric and one or more dimensions for that metric.

Enums

DescribeDimensionKeysError
Errors returned by DescribeDimensionKeys

GetResourceMetricsError
Errors returned by GetResourceMetrics

Traits

PerformanceInsights
Trait representing the capabilities of the AWS PI API. AWS PI clients implement this trait.